fixing liquid_handler import into STAR by generating audio folder + audio.py
