+++ pylabrobot/audio/__init__.py
@@ -1,0 +1,1 @@
+from .audio import *
